pyproject.toml
[tool.poetry]
name = "dummy-service"
version = "0.1.0"
description = "A microservice for MVP"
authors = ["Aditya Sinha <adityasinha2498@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
boto3 = "^1.26.0"
uvicorn = "^0.18.3"
python-dotenv = "^1.0.0"
opentelemetry-api = "^1.30.0"
opentelemetry-sdk = "^1.30.0"
opentelemetry-instrumentation-fastapi = "^0.51b0"
opentelemetry-exporter-otlp = "^1.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
